+++ Assets/Animations/Pedestrians/Ladybug.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: eefb9a693bf094e0d8e35bd0a6025506
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
