+++ Assets/Animations/Pedestrians/Ladybug/Ladybug Crash Left.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 5c1abd0bcf6b5486e98dcbca0d1929af
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
