+++ Assets/Animations/Pedestrians/Ladybug/Ladybug Crash Right.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 52afc2f76fc92479d8fe194075ae43b6
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
