+++ Assets/Animations/Pedestrians/Ladybug/Ladybug Walk E.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 44990b94b78434ef2b5b1198985d7fd8
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
