+++ Assets/Animations/Pedestrians/Ladybug/Ladybug Walk N.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: de1a95b1eec364601a6caf068e06b731
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
