+++ Assets/Animations/Pedestrians/Ladybug/Ladybug Walk S.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 0b628fed7374949e4819fae42825ed76
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
