+++ Assets/Animations/Pedestrians/Ladybug/Ladybug Walk W.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: f5da99a2269154939a00e146138f9a7f
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
