+++ Assets/Animations/Pedestrians/Ladybug/Ladybug.controller.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 13a56d2a620b14ff1bb421d09cf69aab
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 9100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
