+++ Assets/Animations/Pedestrians/RolyPoly.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 71b999185cff4473eb5cea1f616b0274
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
