+++ Assets/Animations/Pedestrians/RolyPoly/Roly Crash Left.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 88c04240ce6cc48ba8a74771d437938b
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
