+++ Assets/Animations/Pedestrians/RolyPoly/Roly Crash Right.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: dad1a6b0bc68a438dba2a8435d4cf789
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
