+++ Assets/Animations/Pedestrians/RolyPoly/Roly Walk E.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: e520c1b7551d541bd862b60606fb47a7
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
