+++ Assets/Animations/Pedestrians/RolyPoly/Roly Walk N.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 4d9f772874add4594b87336889a49c59
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
