+++ Assets/Animations/Pedestrians/RolyPoly/Roly Walk S.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 840112ca1b45742a4a9f5ad809f6a91c
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
