+++ Assets/Animations/Pedestrians/RolyPoly/Roly Walk W.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 93857606d737f469cba39f7d25c17556
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
