+++ Assets/Animations/Pedestrians/RolyPoly/Roly.controller.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: d16c377d639df4edb8fbfabf3a1b6242
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 9100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
