+++ Assets/Art/Sprites/Pedestrians/Ladybug/IMG_0891-ezgif.com-gif-to-sprite-converter.png.meta
@@ -1,0 +1,262 @@
+fileFormatVersion: 2
+guid: c69505d0940d24503b7b312a73f61900
+TextureImporter:
+  internalIDToNameTable: []
+  externalObjects: {}
+  serializedVersion: 13
+  mipmaps:
+    mipMapMode: 0
+    enableMipMap: 0
+    sRGBTexture: 1
+    linearTexture: 0
+    fadeOut: 0
+    borderMipMap: 0
+    mipMapsPreserveCoverage: 0
+    alphaTestReferenceValue: 0.5
+    mipMapFadeDistanceStart: 1
+    mipMapFadeDistanceEnd: 3
+  bumpmap:
+    convertToNormalMap: 0
+    externalNormalMap: 0
+    heightScale: 0.25
+    normalMapFilter: 0
+    flipGreenChannel: 0
+  isReadable: 0
+  streamingMipmaps: 0
+  streamingMipmapsPriority: 0
+  vTOnly: 0
+  ignoreMipmapLimit: 0
+  grayScaleToAlpha: 0
+  generateCubemap: 6
+  cubemapConvolution: 0
+  seamlessCubemap: 0
+  textureFormat: 1
+  maxTextureSize: 2048
+  textureSettings:
+    serializedVersion: 2
+    filterMode: 1
+    aniso: 1
+    mipBias: 0
+    wrapU: 1
+    wrapV: 1
+    wrapW: 0
+  nPOTScale: 0
+  lightmap: 0
+  compressionQuality: 50
+  spriteMode: 2
+  spriteExtrude: 1
+  spriteMeshType: 1
+  alignment: 0
+  spritePivot: {x: 0.5, y: 0.5}
+  spritePixelsToUnits: 100
+  spriteBorder: {x: 0, y: 0, z: 0, w: 0}
+  spriteGenerateFallbackPhysicsShape: 1
+  alphaUsage: 1
+  alphaIsTransparency: 1
+  spriteTessellationDetail: -1
+  textureType: 8
+  textureShape: 1
+  singleChannelComponent: 0
+  flipbookRows: 1
+  flipbookColumns: 1
+  maxTextureSizeSet: 0
+  compressionQualitySet: 0
+  textureFormatSet: 0
+  ignorePngGamma: 0
+  applyGammaDecoding: 0
+  swizzle: 50462976
+  cookieLightType: 0
+  platformSettings:
+  - serializedVersion: 3
+    buildTarget: DefaultTexturePlatform
+    maxTextureSize: 2048
+    resizeAlgorithm: 0
+    textureFormat: -1
+    textureCompression: 1
+    compressionQuality: 50
+    crunchedCompression: 0
+    allowsAlphaSplitting: 0
+    overridden: 0
+    ignorePlatformSupport: 0
+    androidETC2FallbackOverride: 0
+    forceMaximumCompressionQuality_BC6H_BC7: 0
+  - serializedVersion: 3
+    buildTarget: Standalone
+    maxTextureSize: 2048
+    resizeAlgorithm: 0
+    textureFormat: -1
+    textureCompression: 1
+    compressionQuality: 50
+    crunchedCompression: 0
+    allowsAlphaSplitting: 0
+    overridden: 0
+    ignorePlatformSupport: 0
+    androidETC2FallbackOverride: 0
+    forceMaximumCompressionQuality_BC6H_BC7: 0
+  spriteSheet:
+    serializedVersion: 2
+    sprites:
+    - serializedVersion: 2
+      name: IMG_0891-ezgif.com-gif-to-sprite-converter_2
+      rect:
+        serializedVersion: 2
+        x: 2346
+        y: 1972
+        width: 905
+        height: 1705
+      alignment: 0
+      pivot: {x: 0.5, y: 0.5}
+      border: {x: 0, y: 0, z: 0, w: 0}
+      outline: []
+      physicsShape: []
+      tessellationDetail: 0
+      bones: []
+      spriteID: cbe6c396268f04c6795330afd1ff6c55
+      internalID: -1239603184
+      vertices: []
+      indices: 
+      edges: []
+      weights: []
+    - serializedVersion: 2
+      name: IMG_0891-ezgif.com-gif-to-sprite-converter_4
+      rect:
+        serializedVersion: 2
+        x: 415
+        y: 1972
+        width: 914
+        height: 1692
+      alignment: 0
+      pivot: {x: 0.5, y: 0.5}
+      border: {x: 0, y: 0, z: 0, w: 0}
+      outline: []
+      physicsShape: []
+      tessellationDetail: 0
+      bones: []
+      spriteID: fd100d0f5cd8e4aacbf14f3f17445753
+      internalID: -730729141
+      vertices: []
+      indices: 
+      edges: []
+      weights: []
+    - serializedVersion: 2
+      name: IMG_0891-ezgif.com-gif-to-sprite-converter_6
+      rect:
+        serializedVersion: 2
+        x: 4259
+        y: 1958
+        width: 909
+        height: 1718
+      alignment: 0
+      pivot: {x: 0.5, y: 0.5}
+      border: {x: 0, y: 0, z: 0, w: 0}
+      outline: []
+      physicsShape: []
+      tessellationDetail: 0
+      bones: []
+      spriteID: de70c67c5ca6d4247805a23e795454b6
+      internalID: -51000440
+      vertices: []
+      indices: 
+      edges: []
+      weights: []
+    - serializedVersion: 2
+      name: IMG_0891-ezgif.com-gif-to-sprite-converter_7
+      rect:
+        serializedVersion: 2
+        x: 6176
+        y: 1995
+        width: 909
+        height: 1641
+      alignment: 0
+      pivot: {x: 0.5, y: 0.5}
+      border: {x: 0, y: 0, z: 0, w: 0}
+      outline: []
+      physicsShape: []
+      tessellationDetail: 0
+      bones: []
+      spriteID: 3e92b36a0c41f4de5bf37aece59d7fb5
+      internalID: -1053551608
+      vertices: []
+      indices: 
+      edges: []
+      weights: []
+    - serializedVersion: 2
+      name: IMG_0891-ezgif.com-gif-to-sprite-converter_8
+      rect:
+        serializedVersion: 2
+        x: 8102
+        y: 1995
+        width: 905
+        height: 1677
+      alignment: 0
+      pivot: {x: 0.5, y: 0.5}
+      border: {x: 0, y: 0, z: 0, w: 0}
+      outline: []
+      physicsShape: []
+      tessellationDetail: 0
+      bones: []
+      spriteID: 1cde2dcabb48f4f04b56fa6fc17911ac
+      internalID: 1048229470
+      vertices: []
+      indices: 
+      edges: []
+      weights: []
+    - serializedVersion: 2
+      name: IMG_0891-ezgif.com-gif-to-sprite-converter_21
+      rect:
+        serializedVersion: 2
+        x: 429
+        y: 49
+        width: 909
+        height: 1705
+      alignment: 0
+      pivot: {x: 0.5, y: 0.5}
+      border: {x: 0, y: 0, z: 0, w: 0}
+      outline: []
+      physicsShape: []
+      tessellationDetail: 0
+      bones: []
+      spriteID: f2ab360c23d36469a95e2b879dcbea93
+      internalID: 1481228162
+      vertices: []
+      indices: 
+      edges: []
+      weights: []
+    outline: []
+    physicsShape: []
+    bones: []
+    spriteID: 33c4dafbaf7404bf5b2990d0dbbbcecc
+    internalID: 0
+    vertices: []
+    indices: 
+    edges: []
+    weights: []
+    secondaryTextures: []
+    nameFileIdTable:
+      IMG_0891-ezgif.com-gif-to-sprite-converter_0: 232266112
+      IMG_0891-ezgif.com-gif-to-sprite-converter_1: 211362697
+      IMG_0891-ezgif.com-gif-to-sprite-converter_10: -1655540230
+      IMG_0891-ezgif.com-gif-to-sprite-converter_11: -651743735
+      IMG_0891-ezgif.com-gif-to-sprite-converter_12: -1090912560
+      IMG_0891-ezgif.com-gif-to-sprite-converter_13: 885317558
+      IMG_0891-ezgif.com-gif-to-sprite-converter_14: 2115188625
+      IMG_0891-ezgif.com-gif-to-sprite-converter_15: -126855874
+      IMG_0891-ezgif.com-gif-to-sprite-converter_16: -905911712
+      IMG_0891-ezgif.com-gif-to-sprite-converter_17: -1265233901
+      IMG_0891-ezgif.com-gif-to-sprite-converter_18: -811411979
+      IMG_0891-ezgif.com-gif-to-sprite-converter_19: 1683517293
+      IMG_0891-ezgif.com-gif-to-sprite-converter_2: -1239603184
+      IMG_0891-ezgif.com-gif-to-sprite-converter_20: -253035572
+      IMG_0891-ezgif.com-gif-to-sprite-converter_21: 1481228162
+      IMG_0891-ezgif.com-gif-to-sprite-converter_3: -466542711
+      IMG_0891-ezgif.com-gif-to-sprite-converter_4: -730729141
+      IMG_0891-ezgif.com-gif-to-sprite-converter_5: -850484740
+      IMG_0891-ezgif.com-gif-to-sprite-converter_6: -51000440
+      IMG_0891-ezgif.com-gif-to-sprite-converter_7: -1053551608
+      IMG_0891-ezgif.com-gif-to-sprite-converter_8: 1048229470
+      IMG_0891-ezgif.com-gif-to-sprite-converter_9: -1665627669
+  mipmapLimitGroupName: 
+  pSDRemoveMatte: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
