+++ Assets/Scripts/Pedestrian/PedAnimManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a1a4e9643d91a479c91fc539f8995edd
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
